refactor: introduce `ids/idstest`
