+++ ids/idstest/aliases.go
@@ -1,13 +1,14 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-//go:build test
+package idstest
 
-package ids
+import (
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/stretchr/testify/require"
+)
 
-import "github.com/stretchr/testify/require"
-
-var AliasTests = []func(require *require.Assertions, r AliaserReader, w AliaserWriter){
+var AliasTests = []func(require *require.Assertions, r ids.AliaserReader, w ids.AliaserWriter){
 	AliaserLookupErrorTest,
 	AliaserLookupTest,
 	AliaserAliasesEmptyTest,
@@ -17,14 +18,14 @@ var AliasTests = []func(require *require.Assertions, r AliaserReader, w AliaserW
 	AliaserRemoveAliasTest,
 }
 
-func AliaserLookupErrorTest(require *require.Assertions, r AliaserReader, _ AliaserWriter) {
+func AliaserLookupErrorTest(require *require.Assertions, r ids.AliaserReader, _ ids.AliaserWriter) {
 	_, err := r.Lookup("Batman")
 	// TODO: require error to be errNoIDWithAlias
 	require.Error(err) //nolint:forbidigo // currently returns grpc errors too
 }
 
-func AliaserLookupTest(require *require.Assertions, r AliaserReader, w AliaserWriter) {
-	id := ID{'K', 'a', 't', 'e', ' ', 'K', 'a', 'n', 'e'}
+func AliaserLookupTest(require *require.Assertions, r ids.AliaserReader, w ids.AliaserWriter) {
+	id := ids.ID{'K', 'a', 't', 'e', ' ', 'K', 'a', 'n', 'e'}
 	require.NoError(w.Alias(id, "Batwoman"))
 
 	res, err := r.Lookup("Batwoman")
@@ -32,16 +33,16 @@ func AliaserLookupTest(require *require.Assertions, r AliaserReader, w AliaserWr
 	require.Equal(id, res)
 }
 
-func AliaserAliasesEmptyTest(require *require.Assertions, r AliaserReader, _ AliaserWriter) {
-	id := ID{'J', 'a', 'm', 'e', 's', ' ', 'G', 'o', 'r', 'd', 'o', 'n'}
+func AliaserAliasesEmptyTest(require *require.Assertions, r ids.AliaserReader, _ ids.AliaserWriter) {
+	id := ids.ID{'J', 'a', 'm', 'e', 's', ' ', 'G', 'o', 'r', 'd', 'o', 'n'}
 
 	aliases, err := r.Aliases(id)
 	require.NoError(err)
 	require.Empty(aliases)
 }
 
-func AliaserAliasesTest(require *require.Assertions, r AliaserReader, w AliaserWriter) {
-	id := ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
+func AliaserAliasesTest(require *require.Assertions, r ids.AliaserReader, w ids.AliaserWriter) {
+	id := ids.ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
 
 	require.NoError(w.Alias(id, "Batman"))
 	require.NoError(w.Alias(id, "Dark Knight"))
@@ -53,9 +54,9 @@ func AliaserAliasesTest(require *require.Assertions, r AliaserReader, w AliaserW
 	require.Equal(expected, aliases)
 }
 
-func AliaserPrimaryAliasTest(require *require.Assertions, r AliaserReader, w AliaserWriter) {
-	id1 := ID{'J', 'a', 'm', 'e', 's', ' ', 'G', 'o', 'r', 'd', 'o', 'n'}
-	id2 := ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
+func AliaserPrimaryAliasTest(require *require.Assertions, r ids.AliaserReader, w ids.AliaserWriter) {
+	id1 := ids.ID{'J', 'a', 'm', 'e', 's', ' ', 'G', 'o', 'r', 'd', 'o', 'n'}
+	id2 := ids.ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
 
 	require.NoError(w.Alias(id2, "Batman"))
 	require.NoError(w.Alias(id2, "Dark Knight"))
@@ -70,9 +71,9 @@ func AliaserPrimaryAliasTest(require *require.Assertions, r AliaserReader, w Ali
 	require.Equal(expected, res)
 }
 
-func AliaserAliasClashTest(require *require.Assertions, _ AliaserReader, w AliaserWriter) {
-	id1 := ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
-	id2 := ID{'D', 'i', 'c', 'k', ' ', 'G', 'r', 'a', 'y', 's', 'o', 'n'}
+func AliaserAliasClashTest(require *require.Assertions, _ ids.AliaserReader, w ids.AliaserWriter) {
+	id1 := ids.ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
+	id2 := ids.ID{'D', 'i', 'c', 'k', ' ', 'G', 'r', 'a', 'y', 's', 'o', 'n'}
 
 	require.NoError(w.Alias(id1, "Batman"))
 
@@ -81,9 +82,9 @@ func AliaserAliasClashTest(require *require.Assertions, _ AliaserReader, w Alias
 	require.Error(err) //nolint:forbidigo // currently returns grpc errors too
 }
 
-func AliaserRemoveAliasTest(require *require.Assertions, r AliaserReader, w AliaserWriter) {
-	id1 := ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
-	id2 := ID{'J', 'a', 'm', 'e', 's', ' ', 'G', 'o', 'r', 'd', 'o', 'n'}
+func AliaserRemoveAliasTest(require *require.Assertions, r ids.AliaserReader, w ids.AliaserWriter) {
+	id1 := ids.ID{'B', 'r', 'u', 'c', 'e', ' ', 'W', 'a', 'y', 'n', 'e'}
+	id2 := ids.ID{'J', 'a', 'm', 'e', 's', ' ', 'G', 'o', 'r', 'd', 'o', 'n'}
 
 	require.NoError(w.Alias(id1, "Batman"))
 	require.NoError(w.Alias(id1, "Dark Knight"))
